cq.f90: fixed c-binding-type warnings

+++ include/fortran/cq.f90
@@ -31,12 +31,12 @@ end type cq_exec
 ! ------------------------------ HOST OPERATIONS ------------------------------
 abstract interface
   function qkern(NQUBITS, qr, cr, reg) result(status) bind(C)
-    use iso_c_binding, only: c_int
+    use iso_c_binding, only: c_int, c_size_t, c_short
     import :: qubit, qkern_map
     implicit none
-    integer(kind=8), value :: NQUBITS
+    integer(c_size_t), value :: NQUBITS
     type(qubit), value :: qr
-    integer(kind=2) :: cr(0:NQUBITS)
+    integer(c_short) :: cr(0:NQUBITS)
     type(qkern_map), value :: reg
     integer(c_int) :: status
   end function qkern
